Add files via upload
Updating in the correct place. Commit that was previously pushed forward {Added a file for credits, the code needed for to link the home page to credits page and the components in the app file. The credits were modelled closely after the debit implementation so far. Did a little bit of formatting for the home page so the links were more properly formatted. Added the theme of the bank being based in springfield from the simpsons as well.}