structuring for adding debits
Added more things to the add debit function for it to fully work. Added a variable to represent the date the debit was added and added the calculation to take todays date and use it to format the date variable. Created a new debit object that in turn is being added to the array of the debits. Also added quick calculations to add to the sum of debits and subtract from the account balance.